feat: add separate mobile view
